# packages/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Pathfinder API
  description: VEuPathDB Strategy Builder Chatbot API
  version: 1.0.0
paths:
...
  /api/v1/strategies/{strategyId}/steps/{step_id}/analysis-types:
    get:
      tags:
      - steps
      summary: List Analysis Types
      description: List available analysis types for a step.
      operationId: list_analysis_types_api_v1_strategies__strategyId__steps__step_id__analysis_types_get
      security:
      - APIKeyCookie: []
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      - name: step_id
        in: path
        required: true
        schema:
          type: string
          title: Step Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONArray'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/{strategyId}/steps/{step_id}/analysis-types/{analysis_type}:
    get:
      tags:
      - steps
      summary: Get Analysis Type
      description: Get analysis form metadata for a step.
      operationId: get_analysis_type_api_v1_strategies__strategyId__steps__step_id__analysis_types__analysis_type__get
      security:
      - APIKeyCookie: []
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      - name: step_id
        in: path
        required: true
        schema:
          type: string
          title: Step Id
      - name: analysis_type
        in: path
        required: true
        schema:
          type: string
          title: Analysis Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONObject'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/{strategyId}/steps/{step_id}/analyses:
    get:
      tags:
      - steps
      summary: List Step Analyses
      description: List analysis instances for a step.
      operationId: list_step_analyses_api_v1_strategies__strategyId__steps__step_id__analyses_get
      security:
      - APIKeyCookie: []
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      - name: step_id
        in: path
        required: true
        schema:
          type: string
          title: Step Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONArray'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - steps
      summary: Run Step Analysis
      description: Run a step analysis and attach it locally.
      operationId: run_step_analysis_api_v1_strategies__strategyId__steps__step_id__analyses_post
      security:
      - APIKeyCookie: []
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      - name: step_id
        in: path
        required: true
        schema:
          type: string
          title: Step Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepAnalysisRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAnalysisRunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JSONArray:
      items:
        $ref: '#/components/schemas/JSONValue'
      type: array
    JSONObject:
      additionalProperties:
        $ref: '#/components/schemas/JSONValue'
      type: object
    JSONValue: {}
...
    StepAnalysisRequest:
      properties:
        analysisType:
          type: string
          title: Analysistype
        parameters:
          $ref: '#/components/schemas/JSONObject'
        customName:
          anyOf:
          - type: string
          - type: 'null'
          title: Customname
      type: object
      required:
      - analysisType
      title: StepAnalysisRequest
      description: Request to run a step analysis.
    StepAnalysisResponse:
      properties:
        analysisType:
          type: string
          title: Analysistype
        parameters:
          $ref: '#/components/schemas/JSONObject'
        customName:
          anyOf:
          - type: string
          - type: 'null'
          title: Customname
      type: object
      required:
      - analysisType
      title: StepAnalysisResponse
      description: Analysis configuration attached to a step.
    StepAnalysisRunResponse:
      properties:
        analysis:
          $ref: '#/components/schemas/StepAnalysisResponse'
        wdk:
          anyOf:
          - $ref: '#/components/schemas/JSONObject'
          - type: 'null'
      type: object
      required:
      - analysis
      title: StepAnalysisRunResponse
      description: Result of running a step analysis.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: pathfinder-auth
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.pathfinder.veupathdb.org
  description: Production